requirements.txt
langchain
faiss-cpu
openai
